feat(settings): split account controls from app settings
Summary:
- move personal settings into new /account and /account/team pages
- remove user-management destinations from functional navigation
- simplify the session dropdown to user settings and sign out only

Rationale:
- separate operational costing and billing flows from user-specific
  preferences and access management
- give invites and password management a dedicated team-access page
  without exposing user settings in the top app header

Tests:
- npm run typecheck
- git diff --check
- runtime validation via npm run dev with mocked session/user APIs
  for /settings, /account, /account/team, and /settings/users
- npm run lint not run (next lint prompts for initial ESLint setup)
