+++ utils/app-navigation.ts
@@ -10,9 +10,8 @@ export const APP_NAVIGATION_ITEMS: AppNavigationItem[] = [
   { icon: '⊹', children: 'New Purchase Order', href: '/doors/new' },
   { icon: '⊹', children: 'Calculator', href: '/doors/quote' },
   { icon: '⊹', children: 'Customers & Products', href: '/doors/clients' },
-  { icon: '⊹', children: 'Pricing Settings', href: '/settings' },
-  { icon: '⊹', children: 'Billing', href: '/settings/billing' },
-  { icon: '⊹', children: 'User Management', href: '/settings/users' },
+  { icon: '⊹', children: 'Costing Settings', href: '/settings' },
+  { icon: '⊹', children: 'Hosting Billing', href: '/settings/billing' },
   { icon: '⊹', children: 'Component Library', href: '/examples' },
 ];
 
@@ -26,7 +25,6 @@ export const APP_SECTION_ITEMS: AppSectionItem[] = [
   { href: '/doors/new', label: 'New Order' },
   { href: '/doors/quote', label: 'Calculator' },
   { href: '/doors/clients', label: 'Customers' },
-  { href: '/settings', label: 'Pricing' },
+  { href: '/settings', label: 'Costing' },
   { href: '/settings/billing', label: 'Billing' },
-  { href: '/settings/users', label: 'Users' },
 ];
